Updating parsing.
This allows tables with '-<objid>' rows and recognizes (but doesn't
yet handle) '!' updates in tables. '+' updates are also recognized,
but don't change the behavior (they seem to be redundant). It's not
clear that '+' updates are ever actually used in Mork files, but they
show up in the Mork Structure document:

https://developer.mozilla.org/En/Mork_Structure