Re-doing the encoding filter.
Having some fields converted to unicode objects and some not was really
not working well, so now all values are converted. Because of this, the
def-encoding option isn't needed all over the place anymore, so it's
been moved into encoding.py where it's actually used. EncodingStream
no longer needs to do decoding, as DecodeCharacters does it all.